requirements.txt
mysql-connector-python==8.0.25
python-dotenv==0.18.0
twint @ git+https://github.com/twintproject/twint.git@origin/master#egg=twint
